scripts/requirements.txt
azure.core
azure-keyvault-certificates
azure-keyvault-secrets
azure-identity==1.11.0
cryptography
pycryptodome
pyOpenSSL